refactor: XmlMessageSource
