+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/messagesource/XmlMessageSource.java
@@ -1,9 +1,7 @@
-package com.startingblue.fourtooncookie.locale;
+package com.startingblue.fourtooncookie.messagesource;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.support.AbstractMessageSource;
-import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.NodeList;
@@ -16,28 +14,28 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
+import lombok.Getter;
+
 @Slf4j
-@Component
+@Getter
 public class XmlMessageSource extends AbstractMessageSource {
 
-    private static final String DEFAULT_LANGUAGE = "en";
-    private static final String KOREA_LANGUAGE = "ko";
-    private static final String MESSAGE_FILE_PATH_FORMAT = "/messages/messages_%s.xml";
     private static final String ENTRY_TAG_NAME = "entry";
     private static final String KEY_ATTRIBUTE_NAME = "key";
+    private static final String MESSAGE_FILE_PATH_FORMAT = "/messages/messages_%s.xml";
 
     private static final Map<String, Map<String, String>> messages = new ConcurrentHashMap<>();
 
-    public XmlMessageSource() {
-        try {
-            loadMessages(new Locale(DEFAULT_LANGUAGE));
-            loadMessages(new Locale(KOREA_LANGUAGE));
-        } catch (Exception e) {
-            throw new RuntimeException("Failed to load locale messages", e);
-        }
+    private Locale defaultLocale;
+
+    public XmlMessageSource() {}
+
+    public void setDefaultLocale(Locale locale) {
+        this.defaultLocale = locale;
+        setMessages(locale);
     }
 
-    private void loadMessages(Locale locale) throws Exception {
+    public void setMessages(Locale locale) {
         String filename = getMessageFilePath(locale);
         Optional<Document> document = parseXmlDocument(filename);
         if (document.isPresent()) {
@@ -82,13 +80,8 @@ public class XmlMessageSource extends AbstractMessageSource {
     @Override
     public MessageFormat resolveCode(String code, Locale locale) {
         String language = locale.getLanguage();
-        Map<String, String> localeMessages = messages.getOrDefault(language, messages.get(DEFAULT_LANGUAGE));
+        Map<String, String> localeMessages = messages.getOrDefault(language, messages.get(defaultLocale));
         String message = localeMessages.get(code);
-
-        if (!StringUtils.hasText(message)) {
-            throw new IllegalArgumentException(String.format("Message for code '%s' not found for locale '%s'", code, locale));
-        }
-
-        return new MessageFormat(code, locale);
+        return new MessageFormat(message, locale);
     }
 }
